refactor: move Table component to its own file
