+++ src/components/__test__/Table.test.js
@@ -1,23 +1,6 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-
-const Table = ({ children }) => {
-  return (
-    <table>
-      <thead data-testid="thead">
-        <tr>
-          <th>Nome</th>
-          <th>Data de Nascimento</th>
-          <th>Turma</th>
-          <th>Telefone para Emergências</th>
-          <th>Em caso de emergência avisar</th>
-          <th></th>
-        </tr>
-      </thead>
-      <tbody data-testid="tbody">{children}</tbody>
-    </table>
-  );
-};
+import Table from "../Table/Table";
 
 describe("Table component", () => {
   test("should render Table component correctly", () => {
